Moved logic from worker to processor
