+++ snuba/datasets/errors_replacer.py
@@ -56,6 +56,9 @@ In theory this will be needed only during the events to errors migration.
 logger = logging.getLogger(__name__)
 metrics = MetricsWrapper(environment.metrics, "errors.replacer")
 
+CONSUMER_STUFF = "consumer-stuff"
+REPLACER_PROCESSING_TIMEOUT_THRESHOLD = 2 * 60 * 1000  # 2 minutes in ms
+
 
 class ReplacerState(Enum):
     EVENTS = "events"
@@ -485,8 +488,11 @@ class ErrorsReplacer(ReplacerProcessor[Replacement]):
             tag_column_map=self.__tag_column_map,
             promoted_tags=self.__promoted_tags,
         )
+        self.__start_time = float("-inf")
 
     def process_message(self, message: ReplacementMessage) -> Optional[Replacement]:
+        if self._message_already_processed(message):
+            return None
         type_ = message.action_type
 
         if type_ in REPLACEMENT_EVENT_TYPES:
@@ -538,7 +544,45 @@ class ErrorsReplacer(ReplacerProcessor[Replacement]):
 
         return processed
 
+    def _message_already_processed(self, message: ReplacementMessage) -> bool:
+        """
+        Figure out whether or not the message was already processed.
+
+        Check whether there is an entry in Redis for this specific
+        topic, consumer group, and partition. If there is, we can conclude whether
+        or not to process the incoming message by looking at the most recently
+        processed offset.
+        """
+        key = self._build_topic_group_index_key(message.metadata)
+        processed_offset = redis_client.hget(CONSUMER_STUFF, key)
+        if processed_offset is not None:
+            offset = int(processed_offset)
+            return message.metadata.offset <= int(offset)
+
+        return False
+
+    def _build_topic_group_index_key(
+        self, message_metadata: ReplacementMessageMetadata
+    ) -> str:
+        """
+        Builds a unique key for a message being processed for a specific
+        topic, consumer group, and partition.
+
+        Consumer group identifier can be anything that differentiates keys between
+        different consumer groups, it does not necessarily need to be some sort of
+        consumer group ID.
+        """
+        return ":".join(
+            [
+                message_metadata.topic_name,
+                self.__state_name.value,
+                str(message_metadata.partition_index),
+            ]
+        )
+
     def pre_replacement(self, replacement: Replacement, matching_records: int) -> bool:
+
+        self.__start_time = time.time()
 
         # Backward compatibility with the old keys already in Redis, we will let double write
         # the old key structure and the new one for a while then we can get rid of the old one.
@@ -576,6 +620,14 @@ class ErrorsReplacer(ReplacerProcessor[Replacement]):
             return True
 
         return False
+
+    def post_replacement(self, replacement: Replacement, matching_records: int) -> None:
+        if (time.time() - self.__start_time) < REPLACER_PROCESSING_TIMEOUT_THRESHOLD:
+            return
+        message_metadata = replacement.get_message_metadata()
+        key = self._build_topic_group_index_key(message_metadata)
+        redis_client.hset(CONSUMER_STUFF, key, message_metadata.offset + 3)
+        self.__start_time = float("-inf")
 
 
 def _build_event_tombstone_replacement(
